requirements.txt
-i https://pypi.org/simple
certifi==2019.9.11
chardet==3.0.4
click==7.0
flask==1.0.2
future==0.18.1
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
line-bot-sdk==1.14.0
markupsafe==1.1.1
python-dotenv==0.10.3
pygraphviz==1.5
requests==2.22.0
six==1.12.0
transitions==0.6.9
urllib3==1.25.6
werkzeug==0.16.0
beautifulsoup4==4.8.1
pandas==0.25.3
bs4==0.0.1
openpyxl==3.0.5
tweepy==3.10.0
IPython==7.16.1
seaborn==0.11.1
matplotlib==3.3.3 
textblob==0.15.3